Execution enforcer v2/Backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
google-cloud-aiplatform
notion-client
firebase-admin
